refactor: movie lease to pkg/common/leases
